+++ src/accounts/views.py
@@ -1,0 +1,10 @@
+
+from django.contrib.auth.views import LoginView, LogoutView
+
+
+class URLShortenerLoginView(LoginView):
+    template_name = 'accounts/login.html'
+
+
+class URLShortenerLogoutView(LogoutView):
+    pass