+++ CHANGELOG.md
@@ -1,4 +1,19 @@
 # Changelog
 
+# [next] - unreleased
+
+# [0.2.0] - 2023-08-13
+- Added authentication - login page, protected views, logout
+- Added full CRUDLS interface
+  - Update shortened URL
+  - Delete shortened URL
+  - Search for URLs
+  - Pagination
+- Switched to PostgreSQL, add .env support, and improve documentation
+- Demo user support
+- Added tests
+- Add custom error views
+- Setup GitHub workflow which runs unit tests on push
+
 # [0.1.0] - 2023-08-12
 - Initial release. Public URL shortener. Can list and add new URLs.